pyproject.toml
[tool.poetry]
name = "peace-security-test-area-backends"
version = "0.1.0"
description = "平和保障試験区バックエンド群（4 Django プロジェクト + shared）"
readme = "README.md"
authors = ["voidkyun <voidkyun@users.noreply.github.com>"]
packages = [{ include = "shared" }]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.15"
drf-spectacular = "^0.27"
psycopg2-binary = "^2.9"
python-jose = { extras = ["cryptography"], version = "^3.3" }
cryptography = "^43"
pydantic = "^2.0"
requests = "^2.32"
gunicorn = "^23.0"
uvicorn = { extras = ["standard"], version = "^0.32" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.0"

[tool.pytest.ini_options]
testpaths = ["services/root/tests"]
DJANGO_SETTINGS_MODULE = "root.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
